chore: add nonprofit container TODO render nonprofits for container
